Lab 7/requirements.txt
# Lab 7 — Python Dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiofiles==23.2.1
pydantic==2.7.1
python-dotenv==1.0.1
slowapi==0.1.9
# Lab 7 additions
redis==5.0.4
asyncpg==0.29.0
sqlalchemy==2.0.30
aiohttp==3.9.5
requests==2.32.2
# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
